shop/pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["vitaliy <vitalij.leontjev@tetrika.school>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
pydantic = "^2.9.2"
redis = "^5.1.1"
orjson = "^3.10.9"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
fastapi = "^0.115.6"
uvicorn = "^0.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
